requirements.txt
anthropic>=0.40.0
openai>=1.50.0
google-generativeai>=0.8.0
httpx>=0.27.0
bibtexparser>=1.4.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
python-dotenv>=1.0.0
streamlit>=1.35.0
